Refactor schedules into service layer
